+++ patches/src/main/kotlin/app/morphe/patches/reddit/customclients/boostforreddit/http/reddit/UndeleteRedditPatch.kt
@@ -52,6 +52,7 @@ val undeleteRedditPatch = bytecodePatch(
             addInstructions(
                 index,
                 """
+                invoke-static       { }, $OKHTTP_EXTENSION_CLASS_DESCRIPTOR->init()V
                 invoke-static       { v0 }, $OKHTTP_EXTENSION_CLASS_DESCRIPTOR->installInterceptor(Lokhttp3/OkHttpClient${'$'}Builder;)Lokhttp3/OkHttpClient${'$'}Builder;
                 move-result-object  v0
                 """